requirements.txt
gevent==1.1.1
Jinja2==2.8
requests==2.9.1
PyYAML==3.11